ai-agent-langchain/requirements.txt
langchain
langchain-community
langchain_openai
yfinance
python-dotenv
faiss-cpu
pypdf
duckduckgo-search